08-local-agent-runtime/backend/requirements.txt
# Local Agent Runtime - December 2025 Stack

# LangGraph (production-ready orchestration)
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0

# Ollama (local LLM with function calling)
langchain-ollama>=0.2.0

# MCP (Model Context Protocol - November 2025 spec)
mcp>=1.0.0
httpx>=0.27.0

# FastAPI (for MCP servers)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Utilities
pyyaml>=6.0
pydantic>=2.5.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

